Relocate Creality E3V2 files
