+++ Marlin/src/gcode/bedlevel/abl/G29.cpp
@@ -61,7 +61,7 @@
 #endif
 
 #if ENABLED(DWIN_CREALITY_LCD)
-  #include "../../../lcd/dwin/e3v2/dwin.h"
+  #include "../../../lcd/e3v2/creality/dwin.h"
 #endif
 
 #if HAS_MULTI_HOTEND
