+++ Marlin/src/lcd/e3v2/creality/dwin.h
@@ -25,7 +25,7 @@
  * DWIN by Creality3D
  */
 
-#include "../dwin_lcd.h"
+#include "dwin_lcd.h"
 #include "rotary_encoder.h"
 #include "../../../libs/BL24CXX.h"
