+++ Marlin/src/lcd/e3v2/creality/rotary_encoder.cpp
@@ -21,7 +21,7 @@
  */
 
 /*****************************************************************************
- * @file     rotary_encoder.cpp
+ * @file     lcd/e3v2/creality/rotary_encoder.cpp
  * @author   LEO / Creality3D
  * @date     2019/07/06
  * @version  2.0.1
