+++ ini/features.ini
@@ -44,8 +44,8 @@ HAS_SPI_TFT                            = src_filter=+<src/HAL/STM32/tft/tft_spi.
 I2C_EEPROM                             = src_filter=+<src/HAL/shared/eeprom_if_i2c.cpp>
 SOFT_I2C_EEPROM                        = SlowSoftI2CMaster, SlowSoftWire=https://github.com/felias-fogg/SlowSoftWire/archive/master.zip
 SPI_EEPROM                             = src_filter=+<src/HAL/shared/eeprom_if_spi.cpp>
+DWIN_CREALITY_LCD                      = src_filter=+<src/lcd/e3v2/creality>
 HAS_GRAPHICAL_TFT                      = src_filter=+<src/lcd/tft>
-DWIN_CREALITY_LCD                      = src_filter=+<src/lcd/dwin>
 IS_TFTGLCD_PANEL                       = src_filter=+<src/lcd/TFTGLCD>
 HAS_TOUCH_BUTTONS                      = src_filter=+<src/lcd/touch/touch_buttons.cpp>
 HAS_LCD_MENU                           = src_filter=+<src/lcd/menu>
